refactor: Organizar classes dos ViewComponents
