+++ BaseDeProjetos/BaseDeProjetos.csproj
@@ -28,9 +28,6 @@
     <PackageReference Include="SendGrid" Version="9.21.0" />
   </ItemGroup>
 
-  <ItemGroup>
-  </ItemGroup>
-
   <ProjectExtensions><VisualStudio><UserProperties appsettings_1json__JsonSchema="" /></VisualStudio></ProjectExtensions>
 
 </Project>
