+++ BaseDeProjetos/ViewComponents/EditalViewComponents/ModalEditEditalViewComponent.cs
@@ -1,21 +1,22 @@
 ﻿using BaseDeProjetos.Data;
 using Microsoft.AspNetCore.Mvc;
+using System;
 using System.Threading.Tasks;
 
-namespace BaseDeProjetos.ViewComponents
+namespace BaseDeProjetos.ViewComponents.EditalViewComponents
 {
-    public class ModalHistoryProspViewComponent : ViewComponent
+    public class ModalEditEditalViewComponent : ViewComponent
     {
         private readonly ApplicationDbContext _context;
 
-        public ModalHistoryProspViewComponent(ApplicationDbContext context)
+        public ModalEditEditalViewComponent(ApplicationDbContext context)
         {
             _context = context;
         }
 
         public async Task<IViewComponentResult> InvokeAsync(string id)
         {
-            var model = await _context.Prospeccao.FindAsync(id);
+            var model = await _context.Editais.FindAsync(int.Parse(id));
             return View(model);
         }
     }
