+++ BaseDeProjetos/ViewComponents/FunilDeVendasViewComponents/ModalDeleteProspViewComponent.cs
@@ -2,20 +2,20 @@
 using Microsoft.AspNetCore.Mvc;
 using System.Threading.Tasks;
 
-namespace BaseDeProjetos.ViewComponents
+namespace BaseDeProjetos.ViewComponents.FunilDeVendasViewComponents
 {
-    public class ModalDeleteProjetoViewComponent : ViewComponent
+    public class ModalDeleteProspViewComponent : ViewComponent
     {
         private readonly ApplicationDbContext _context;
 
-        public ModalDeleteProjetoViewComponent(ApplicationDbContext context)
+        public ModalDeleteProspViewComponent(ApplicationDbContext context)
         {
             _context = context;
         }
 
         public async Task<IViewComponentResult> InvokeAsync(string id)
         {
-            var model = await _context.Projeto.FindAsync(id);
+            var model = await _context.Prospeccao.FindAsync(id);
             return View(model);
         }
     }
