+++ BaseDeProjetos/ViewComponents/FunilDeVendasViewComponents/ModalEditProspViewComponent.cs
@@ -2,21 +2,20 @@
 using Microsoft.AspNetCore.Mvc;
 using System.Threading.Tasks;
 
-namespace BaseDeProjetos.ViewComponents
+namespace BaseDeProjetos.ViewComponents.FunilDeVendasViewComponents
 {
-    public class ModalDetalhesProjetoViewComponent : ViewComponent
-
+    public class ModalEditProspViewComponent : ViewComponent
     {
         private readonly ApplicationDbContext _context;
 
-        public ModalDetalhesProjetoViewComponent(ApplicationDbContext context)
+        public ModalEditProspViewComponent(ApplicationDbContext context)
         {
             _context = context;
         }
 
         public async Task<IViewComponentResult> InvokeAsync(string id)
         {
-            var model = await _context.Projeto.FindAsync(id);
+            var model = await _context.Prospeccao.FindAsync(id);
             return View(model);
         }
     }
