+++ BaseDeProjetos/ViewComponents/FunilDeVendasViewComponents/ModalPreCreateProspViewComponent.cs
@@ -2,14 +2,16 @@
 using Microsoft.AspNetCore.Mvc;
 using System.Threading.Tasks;
 using BaseDeProjetos.Models;
+using System.Linq;
+using System.Collections.Generic;
 
-namespace BaseDeProjetos.ViewComponents
+namespace BaseDeProjetos.ViewComponents.FunilDeVendasViewComponents
 {
-    public class ModalCreateProspViewComponent : ViewComponent
+    public class ModalPreCreateProspViewComponent : ViewComponent
     {
         private readonly ApplicationDbContext _context;
 
-        public ModalCreateProspViewComponent(ApplicationDbContext context)
+        public ModalPreCreateProspViewComponent(ApplicationDbContext context)
         {
             _context = context;
         }
@@ -17,6 +19,10 @@ namespace BaseDeProjetos.ViewComponents
 
         public async Task<IViewComponentResult> InvokeAsync()
         {
+            List<Prospeccao> prosp = _context.Prospeccao.ToList();
+
+            ViewData["prospPlan"] = prosp;
+
             return View(new Prospeccao(new FollowUp()));
         }
     }
