--- BaseDeProjetos/ViewComponents/ModalCreateProjetoViewComponent.cs
@@ -1,15 +1,0 @@
-﻿using BaseDeProjetos.Data;
-using BaseDeProjetos.Models;
-using Microsoft.AspNetCore.Mvc;
-using System.Threading.Tasks;
-
-namespace BaseDeProjetos.ViewComponents
-{
-	public class ModalCreateProjetoViewComponent : ViewComponent
-	{
-		public async Task<IViewComponentResult> InvokeAsync()
-		{
-			return View(new Projeto());
-		}
-	}
-}
