--- BaseDeProjetos/ViewComponents/ModalDeleteProspViewComponent.cs
@@ -1,22 +1,0 @@
-﻿using BaseDeProjetos.Data;
-using Microsoft.AspNetCore.Mvc;
-using System.Threading.Tasks;
-
-namespace BaseDeProjetos.ViewComponents
-{
-	public class ModalDeleteProspViewComponent : ViewComponent
-	{
-		private readonly ApplicationDbContext _context;
-
-		public ModalDeleteProspViewComponent(ApplicationDbContext context)
-		{
-			_context = context;
-		}
-
-		public async Task<IViewComponentResult> InvokeAsync(string id)
-		{
-			var model = await _context.Prospeccao.FindAsync(id);
-			return View(model);
-		}
-	}
-}
