+++ BaseDeProjetos/ViewComponents/ProducaoViewComponents/ModalDeleteProducaoViewComponent.cs
@@ -2,7 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using System.Threading.Tasks;
 
-namespace BaseDeProjetos.ViewComponents
+namespace BaseDeProjetos.ViewComponents.ProducaoViewComponents
 {
     public class ModalDeleteProducaoViewComponent : ViewComponent
     {
