+++ BaseDeProjetos/ViewComponents/ProjetoViewComponents/ModalCreateProjetoViewComponent.cs
@@ -1,0 +1,15 @@
+﻿using BaseDeProjetos.Data;
+using BaseDeProjetos.Models;
+using Microsoft.AspNetCore.Mvc;
+using System.Threading.Tasks;
+
+namespace BaseDeProjetos.ViewComponents.ProjetoViewComponents
+{
+    public class ModalCreateProjetoViewComponent : ViewComponent
+    {
+        public async Task<IViewComponentResult> InvokeAsync()
+        {
+            return View(new Projeto());
+        }
+    }
+}
