+++ BaseDeProjetos/ViewComponents/ProjetoViewComponents/ModalDetalhesProjetoViewComponent.cs
@@ -1,22 +1,22 @@
 ﻿using BaseDeProjetos.Data;
 using Microsoft.AspNetCore.Mvc;
-using System;
 using System.Threading.Tasks;
 
-namespace BaseDeProjetos.ViewComponents
+namespace BaseDeProjetos.ViewComponents.ProjetoViewComponents
 {
-    public class ModalEditEditalViewComponent : ViewComponent
+    public class ModalDetalhesProjetoViewComponent : ViewComponent
+
     {
         private readonly ApplicationDbContext _context;
 
-        public ModalEditEditalViewComponent(ApplicationDbContext context)
+        public ModalDetalhesProjetoViewComponent(ApplicationDbContext context)
         {
             _context = context;
         }
 
         public async Task<IViewComponentResult> InvokeAsync(string id)
-        { 
-            var model = await _context.Editais.FindAsync(Int32.Parse(id));
+        {
+            var model = await _context.Projeto.FindAsync(id);
             return View(model);
         }
     }
